requirements.txt
jinja2
openai
requests
# token counting helper used when available
tiktoken
# optional: vllm (for local model hosting)
# vllm
# YAML parsing for config
pyyaml
